Refactored `_Internal` package: moved AWS-specific modules to `Connect/_Internal`, streamlined imports, removed `ccLogBase` dependency, and integrated `__init__.py` for centralized internal utility management.
