+++ WrenchCL/Connect/_Internal/_boto_cache.py
@@ -3,18 +3,15 @@
 #  Licensed under the MIT License (https://opensource.org/license/mit).
 
 from functools import lru_cache
-from typing import Optional, TYPE_CHECKING, Any
-
-if TYPE_CHECKING:
-    pass
+from typing import Optional, Any
 
 try:
-    from botocore.config import Config as RuntimeConfig
-    import boto3 as runtime_boto3
+    from botocore.config import Config
+    import boto3
 
     @lru_cache(maxsize=3)
-    def _get_boto3_session(profile_name: str) -> Any:
-        return runtime_boto3.session.Session(profile_name=profile_name)
+    def _get_boto3_session(profile_name: str) -> Any:  # Use Any instead of boto3.Session
+        return boto3.session.Session(profile_name=profile_name)
 
     @lru_cache(maxsize=6)
     def _fetch_secret_from_secretsmanager(profile: str, region: str, secret_arn: str) -> str:
@@ -22,7 +19,7 @@ try:
         return client.get_secret_value(SecretId=secret_arn)['SecretString']
 
     @lru_cache(maxsize=3)
-    def _get_s3_client(profile: str, region: str, config: Optional[RuntimeConfig] = None) -> Any:
+    def _get_s3_client(profile: str, region: str, config: Optional[Config] = None) -> Any:
         client = _get_boto3_session(profile).client(service_name='s3', region_name=region, config=config)
         return client
 
@@ -30,4 +27,4 @@ except ImportError:
     _get_boto3_session = None
     _fetch_secret_from_secretsmanager = None
     _get_s3_client = None
-    RuntimeConfig = None
+    Config = None
